modified:   docs/config.md 	modified:   sendit/apps/main/models.py 	modified:   sendit/apps/main/tasks.py 	modified:   sendit/apps/main/utils.py 	modified:   sendit/settings/config.py

+++ sendit/apps/main/tasks.py
@@ -44,19 +44,22 @@ from sendit.apps.main.utils import (
 
 from som.api.identifiers.dicom import (
     get_identifiers as get_ids,
-    replace_identifiers as replace_ids
+    replace_identifiers as replace_ids,
+    prepare_identifiers
 )
 
 from som.api.identifiers import Client
 
 from sendit.settings import (
     DEIDENTIFY_RESTFUL,
+    DEIDENTIFY_PIXELS,
     SEND_TO_ORTHANC,
     SOM_STUDY,
     ORTHANC_IPADDRESS,
     ORTHANC_PORT,
     SEND_TO_GOOGLE,
-    GOOGLE_CLOUD_STORAGE
+    GOOGLE_CLOUD_STORAGE,
+    GOOGLE_STORAGE_COLLECTION
 )
 
 from django.conf import settings
@@ -79,6 +82,10 @@ app.autodiscover_tasks(lambda: settings.INSTALLED_APPS)
 
 @shared_task
 def import_dicomdir(dicom_dir):
+    '''import dicom directory manages importing a valid dicom set into the application,
+    and is a celery job triggered by the watcher. Here we also flag (and disclude)
+    images that have a header value that indicates pixel identifiers.
+    '''
     if os.path.exists(dicom_dir):
         dicom_files = ls_fullpath(dicom_dir)
         bot.debug("Importing %s, found %s .dcm files" %(dicom_dir,len(dicom_files)))        
@@ -91,6 +98,7 @@ def import_dicomdir(dicom_dir):
         # Add in each dicom file to the series
         for dcm_file in dicom_files:
             try:
+
                 # The dicom folder will be named based on the accession#
                 dcm = read_file(dcm_file,force=True)
                 if patient_id is None:
@@ -98,18 +106,25 @@ def import_dicomdir(dicom_dir):
 
                 # if different patient flag as erroneous
                 if patient_id == dcm.PatientID:
-
                     dicom_uid = os.path.basename(dcm_file)
 
-                    # Create the Image object in the database
-                    # A dicom instance number must be unique for its batch
-                    dicom = Image.objects.create(batch=batch,
-                                             uid=dicom_uid)
+                    # If the image has pixel identifiers, we don't include 
+                    if dcm.get("BurnedInAnnotation") is not None:
+                        message = "%s has burned pixel annotation, skipping" %dicom_uid
+                        batch = add_batch_error(message,batch)
 
-                    # Save the dicom file to storage
-                    dicom = save_image_dicom(dicom=dicom,
-                                             dicom_file=dcm_file) # Also saves
-                    os.remove(dcm_file)
+                    else:
+                        # Create the Image object in the database
+                        # A dicom instance number must be unique for its batch
+                        dicom = Image.objects.create(batch=batch,
+                                                     uid=dicom_uid)
+
+                        # Save the dicom file to storage
+                        dicom = save_image_dicom(dicom=dicom,
+                                                 dicom_file=dcm_file) # Also saves
+
+                        # Only remove files successfully imported
+                        os.remove(dcm_file)
                 else:
                     message = "Mismatch PatientID for file %s" %(dcm_file)
                     batch = add_batch_error(message,batch)
@@ -128,7 +143,7 @@ def import_dicomdir(dicom_dir):
         # If there were no errors on import, we should remove the directory
         if not batch.has_error:
 
-            # Should trigger error if not empty
+            # Should only be called given no error, and should trigger error if not empty
             os.rmdir(dicom_dir)
 
         # At the end, submit the dicoms to be deidentified as a batch 
@@ -136,10 +151,48 @@ def import_dicomdir(dicom_dir):
         if count > 0:
             bot.debug("Submitting task to get_identifiers for batch %s with %s dicoms." %(batch.uid,
                                                                                           count))
+            if DEIDENTIFY_PIXELS is True:
+                bot.warning("Deidentify pixels is not yet implemented. Images were skipped.")
+                # When this is implemented, the function will be modified to add these images
+                # to the batch, which will then be first sent through a function to
+                # scrub pixels before header data is looked at.
+                # scrub_pixels.apply_async(kwargs={"bid":batch.id})
+            #else:
             get_identifiers.apply_async(kwargs={"bid":batch.id})
 
     else:
         bot.warning('Cannot find %s' %dicom_dir)
+
+
+@shared_task
+def scrub_pixels(bid):
+    '''scrub pixels (not currently triggered) will be run to scrub pixel data
+    before identifiers are extracted from the header.
+    '''
+    batch = Batch.objects.get(id=bid)
+    images = batch.image_set.all()
+    dicom_files = batch.get_image_paths()
+    batch.change_images_status('PROCESSING')
+
+    # from deid.dicom import scrub_pixels
+
+    for dcm_file in dicom_files:
+
+        dicom_uid = os.path.basename(dcm_file)
+
+        if dicom.get("BurnedInAnnotation") is not None:
+
+            # We shouldn't be in this function if False, but we will check again anyway
+            if DEIDENTIFY_PIXELS is True:
+                print("De-identification will be done here.")
+            else:
+                #TODO: remove image from batch here, not okay to upload
+                message = "%s has pixel identifiers, deidentify pixels is off, but added to batch. Removing!" %dicom_uid
+                batch = add_batch_error(message,batch)
+
+    # At the end, move on to processing headers    
+    get_identifiers.apply_async(kwargs={"bid":batch.id})
+ 
 
 
 @shared_task
@@ -170,17 +223,17 @@ def get_identifiers(bid,study=None):
         ids = get_ids(dicom_files=dicom_files)
 
         bot.debug("som.client making request to deidentify batch %s" %(bid))
-        #:param save_records: if True, will use mrn endpoint and and save data.
-        #                     if False, will use uid endpoint and not save data.
-        # Right now save_records is set to False, not sure if it should be
-        result = cli.deidentify(ids=ids,study=study)     # should return a list
-          
-        batch_ids = BatchIdentifiers.objects.create(batch=batch,
-                                                    response=result)
-        batch_ids.save()
-        
-        replace_identifiers.apply_async(kwargs={"bid":bid})
+        result = cli.deidentify(ids=ids,study=study)  # should return dict with "results"
 
+        if "results" in result:                  
+            batch_ids = BatchIdentifiers.objects.create(batch=batch,
+                                                        response=result['results'])
+            batch_ids.save()        
+            replace_identifiers.apply_async(kwargs={"bid":bid})
+        else:
+            message = "Error calling som uid endpoint: %s" %result
+            batch = add_batch_error(message,batch)
+        
 
     else:
         bot.debug("Restful de-identification skipped [DEIDENTIFY_RESTFUL is False]")
@@ -205,10 +258,22 @@ def replace_identifiers(bid):
         # We are going to overwrite, they have same base
         output_folder = os.path.dirname(dicom_files[0])
 
+        # Prepare the identifiers - not this isn't necessary, but we do it
+        # so we can rename the images before upload
+        ids = prepare_identifiers(response=batch_ids.response,
+                                  dicom_files=dicom_files)
+
+        # Now we have a lookup with ids[entity_id][field]
+        for dcm in batch.image_set.all():
+            dicom = dcm.load_dicom()
+            # STOPPED HERE - need to rename dicom with suid
+
         updated_files = replace_ids(dicom_files=dicom_files,
                                     response=batch_ids.response,
                                     overwrite=True,
                                     output_folder=output_folder)  
+
+        DEIDENTIFY_PIXELS=False
 
         change_status(batch,"DONEPROCESSING")
         batch.change_images_status('DONEPROCESSING')
@@ -238,8 +303,30 @@ def upload_storage(bid):
         # do the send here!
 
     if SEND_TO_GOOGLE is True and GOOGLE_CLOUD_STORAGE not in [None,""]:
-        bot.log("Uploading to Google Storage %s" %(GOOGLE_CLOUD_STORAGE))
-        # GOOGLE_CLOUD_STORAGE
+
+        if GOOGLE_STORAGE_COLLECTION is not None:
+            from som.api.google.storage import Client
+            bot.log("Uploading to Google Storage %s" %(GOOGLE_CLOUD_STORAGE))
+
+            # Question: what fields to include as metadata?
+            # all in header (this includes image dimensions)
+            # study
+
+            # PREPARE FILES HERE
+            # updated files...
+
+            # Upload the dataset
+            client.upload_dataset(images=updated_files,
+                                  collection=collection,
+                                  uid=metadata['id'],
+                                  entity_metadata=metadata)
+  
+            client = Client(bucket_name=GOOGLE_CLOUD_STORAGE)
+            collection = client.create_collection(uid=SOM_STUDY)
+
+        else:
+            message = "batch %s send to Google skipped, no storage collection defined." %batch
+            batch = add_batch_error(message,batch)
 
         batch.change_images_status('SENT')
